LLM_Twin/requirements.txt
# Core dependencies
transformers==4.36.2
torch==2.1.2
datasets==2.14.6
accelerate==0.25.0
bitsandbytes==0.41.1
sentence-transformers==2.2.2

# Data Collection
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
fastapi==0.105.0
uvicorn==0.24.0

# Database
pymongo==4.6.1
qdrant-client==1.6.4
faiss-cpu==1.7.4

# Web Interface
streamlit==1.28.2
python-dotenv==1.0.0

# Utilities
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1 